+++ web/src/view/superAdmin/api/api.vue
@@ -1,34 +1,6 @@
 <template>
   <div>
-    <div class="search-term">
-      <el-form :inline="true" :model="searchInfo" class="demo-form-inline">
-        <el-form-item label="路径">
-          <el-input placeholder="路径" v-model="searchInfo.path" @keyup.enter.native="onSubmit"></el-input>
-        </el-form-item>
-        <el-form-item label="描述">
-          <el-input placeholder="描述" v-model="searchInfo.description" @keyup.enter.native="onSubmit"></el-input>
-        </el-form-item>
-        <el-form-item label="api组">
-          <el-input placeholder="api组" v-model="searchInfo.apiGroup" @keyup.enter.native="onSubmit"></el-input>
-        </el-form-item>
-        <el-form-item label="请求">
-          <el-select clearable placeholder="请选择" v-model="searchInfo.method" @change="onSubmit">
-            <el-option
-              :key="item.value"
-              :label="`${item.label}(${item.value})`"
-              :value="item.value"
-              v-for="item in methodOptions"
-            ></el-option>
-          </el-select>
-        </el-form-item>
-        <el-form-item>
-          <el-button @click="onReset" type="primary">重置</el-button>
-        </el-form-item>
-        <el-form-item>
-          <el-button @click="openDialog('addApi')" type="primary">新增api</el-button>
-        </el-form-item>
-      </el-form>
-    </div>
+    <search-form @change="handleFormChange" @add="$refs.dialog.openDialog('addApi')" />
 
     <pagination-table ref="table" :dataSource="getTableData">
       <el-table slot-scope="data" :data="data.tableData" @sort-change="sortChange" border stripe>
@@ -40,121 +12,63 @@
           <template slot-scope="scope">
             <div>
               {{ scope.row.method }}
-              <el-tag
-                :key="scope.row.methodFiletr"
-                :type="scope.row.method | tagTypeFiletr"
-                effect="dark"
-                size="mini"
-                >{{ scope.row.method | methodFiletr }}</el-tag
-              >
-              <!-- {{scope.row.method|methodFiletr}} -->
+              <el-tag :key="scope.row.methodFiletr" :type="scope.row.method | tagTypeFiletr" effect="dark" size="mini">
+                {{ scope.row.method | methodFiletr }}
+              </el-tag>
             </div>
           </template>
         </el-table-column>
-
         <el-table-column fixed="right" label="操作" width="200">
           <template slot-scope="scope">
-            <el-button @click="editApi(scope.row)" size="small" type="primary" icon="el-icon-edit">编辑</el-button>
+            <el-button
+              @click="$refs.dialog.openDialog('edit', scope.row)"
+              size="small"
+              type="primary"
+              icon="el-icon-edit"
+              >编辑</el-button
+            >
             <el-button @click="deleteApi(scope.row)" size="small" type="danger" icon="el-icon-delete">删除</el-button>
           </template>
         </el-table-column>
       </el-table>
     </pagination-table>
 
-    <el-dialog :before-close="closeDialog" :title="dialogTitle" :visible.sync="dialogFormVisible">
-      <el-form :inline="true" :model="form" :rules="rules" label-width="80px" ref="apiForm">
-        <el-form-item label="路径" prop="path">
-          <el-input autocomplete="off" v-model="form.path"></el-input>
-        </el-form-item>
-        <el-form-item label="请求" prop="method">
-          <el-select placeholder="请选择" v-model="form.method">
-            <el-option
-              :key="item.value"
-              :label="`${item.label}(${item.value})`"
-              :value="item.value"
-              v-for="item in methodOptions"
-            ></el-option>
-          </el-select>
-        </el-form-item>
-        <el-form-item label="api分组" prop="apiGroup">
-          <el-input autocomplete="off" v-model="form.apiGroup"></el-input>
-        </el-form-item>
-        <el-form-item label="api简介" prop="description">
-          <el-input autocomplete="off" v-model="form.description"></el-input>
-        </el-form-item>
-      </el-form>
-      <div class="warning">新增Api需要在角色管理内配置权限才可使用</div>
-      <div class="dialog-footer" slot="footer">
-        <el-button @click="closeDialog">取 消</el-button>
-        <el-button @click="enterDialog" type="primary">确 定</el-button>
-      </div>
-    </el-dialog>
+    <Dialog ref="dialog" />
   </div>
 </template>
 
 <script>
 // 获取列表内容封装在mixins内部  getTableData方法 初始化已封装完成 条件搜索时候 请把条件安好后台定制的结构体字段 放到 this.searchInfo 中即可实现条件搜索
 
-import { getApiById, getApiList, createApi, updateApi, deleteApi } from '@/api/api'
-import PaginationTable from '@/components/PaginationTable'
+import { getApiList, deleteApi } from '@/api/api'
 import { toSQLLine } from '@/utils/stringFun'
-const methodOptions = [
-  {
-    value: 'POST',
-    label: '创建',
-    type: 'success',
-  },
-  {
-    value: 'GET',
-    label: '查看',
-    type: '',
-  },
-  {
-    value: 'PUT',
-    label: '更新',
-    type: 'warning',
-  },
-  {
-    value: 'DELETE',
-    label: '删除',
-    type: 'danger',
-  },
-]
+
+import PaginationTable from '@/components/PaginationTable'
+import SearchForm from './components/SearchForm'
+import Dialog from './components/Dialog'
+
+import config from './components/config'
 
 export default {
   name: 'Api',
   components: {
     PaginationTable,
+    SearchForm,
+    Dialog,
   },
   data() {
     return {
+      methodOptions: config.methodOptions,
       searchInfo: {},
-      dialogFormVisible: false,
-      dialogTitle: '新增Api',
-      form: {
-        path: '',
-        apiGroup: '',
-        method: '',
-        description: '',
-      },
-      methodOptions: methodOptions,
-      type: '',
-      rules: {
-        path: [{ required: true, message: '请输入api路径', trigger: 'blur' }],
-        apiGroup: [{ required: true, message: '请输入组名称', trigger: 'blur' }],
-        method: [{ required: true, message: '请选择请求方式', trigger: 'blur' }],
-        description: [{ required: true, message: '请输入api介绍', trigger: 'blur' }],
-      },
     }
   },
   filters: {
     methodFiletr(value) {
-      const target = methodOptions.filter(item => item.value === value)[0]
-      // return target && `${target.label}(${target.value})`
+      const target = config.methodOptions.filter(item => item.value === value)[0]
       return target && `${target.label}`
     },
     tagTypeFiletr(value) {
-      const target = methodOptions.filter(item => item.value === value)[0]
+      const target = config.methodOptions.filter(item => item.value === value)[0]
       return target && `${target.type}`
     },
   },
@@ -162,6 +76,11 @@ export default {
     // 通过此方法将搜索参数传入组件
     getTableData(params) {
       return getApiList({ ...params, ...this.searchInfo })
+    },
+    //  搜索条件变换后执行
+    handleFormChange(searchInfo) {
+      this.searchInfo = searchInfo
+      this.$refs.table.getTableData()
     },
     // 排序
     sortChange({ prop, order }) {
@@ -171,47 +90,7 @@ export default {
       }
       this.$refs.table.getTableData()
     },
-    //条件搜索前端看此方法
-    onSubmit() {
-      this.$refs.table.getTableData()
-    },
-    // 重置搜索条件
-    onReset() {
-      this.searchInfo = {}
-      this.$refs.table.changePage(1)
-    },
-    initForm() {
-      this.$refs.apiForm.resetFields()
-      this.form = {
-        path: '',
-        apiGroup: '',
-        method: '',
-        description: '',
-      }
-    },
-    closeDialog() {
-      this.initForm()
-      this.dialogFormVisible = false
-    },
-    openDialog(type) {
-      switch (type) {
-        case 'addApi':
-          this.dialogTitlethis = '新增Api'
-          break
-        case 'edit':
-          this.dialogTitlethis = '编辑Api'
-          break
-        default:
-          break
-      }
-      this.type = type
-      this.dialogFormVisible = true
-    },
-    async editApi(row) {
-      const res = await getApiById({ id: row.ID })
-      this.form = res.data.api
-      this.openDialog('edit')
-    },
+    // 删除
     async deleteApi(row) {
       this.$confirm('此操作将永久删除所有角色下该菜单, 是否继续?', '提示', {
         confirmButtonText: '确定',
@@ -234,52 +113,6 @@ export default {
             message: '已取消删除',
           })
         })
-    },
-    async enterDialog() {
-      this.$refs.apiForm.validate(async valid => {
-        if (valid) {
-          switch (this.type) {
-            case 'addApi':
-              {
-                const res = await createApi(this.form)
-                if (res.code == 0) {
-                  this.$message({
-                    type: 'success',
-                    message: '添加成功',
-                    showClose: true,
-                  })
-                }
-                this.$refs.table.getTableData()
-                this.closeDialog()
-              }
-
-              break
-            case 'edit':
-              {
-                const res = await updateApi(this.form)
-                if (res.code == 0) {
-                  this.$message({
-                    type: 'success',
-                    message: '编辑成功',
-                    showClose: true,
-                  })
-                }
-                this.$refs.table.getTableData()
-                this.closeDialog()
-              }
-              break
-            default:
-              {
-                this.$message({
-                  type: 'error',
-                  message: '未知操作',
-                  showClose: true,
-                })
-              }
-              break
-          }
-        }
-      })
     },
   },
 }
