+++ web/src/view/superAdmin/api/components/config.js
@@ -1,0 +1,24 @@
+export default {
+  methodOptions: [
+    {
+      value: 'POST',
+      label: '创建',
+      type: 'success',
+    },
+    {
+      value: 'GET',
+      label: '查看',
+      type: '',
+    },
+    {
+      value: 'PUT',
+      label: '更新',
+      type: 'warning',
+    },
+    {
+      value: 'DELETE',
+      label: '删除',
+      type: 'danger',
+    },
+  ],
+}
